MOD structure w/ headers and sources separated
