+++ include/proper_primes.hpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 #define PRIMES_NO 5483
-uint32_t proper_primes[5483] = {
+const uint32_t proper_primes[5483] = {
 3, 
 5, 
 11,
