+++ src/constant_weight_encodable_bits.cpp
@@ -12,9 +12,9 @@ int main(int argc, char* argv[]){
     return -1;
  }
 
+ InitConstants();
  InitBinomials();
  NTL::RR::SetPrecision(NUM_BITS_REAL_MANTISSA);
- pi = NTL::ComputePi_RR();
  uint32_t n = atoi(argv[1]);
  uint32_t t = atoi(argv[2]);
  /* reduce by a factor matching the QC block size */
