+++ src/parameter_generator.cpp
@@ -5,9 +5,9 @@
 
 #define NUM_BITS_REAL_MANTISSA 128
 #define IGNORE_DECODING_COST 0
-#define SKIP_BJMM 1
-#define SKIP_MMT 1
-#define LOG_COST_CRITERION 1
+// #define SKIP_BJMM 1
+// #define SKIP_MMT 1
+// #define LOG_COST_CRITERION 1
 
 #include "binomials.hpp"
 #include "bit_error_probabilities.hpp"
@@ -160,6 +160,7 @@ int main(int argc, char *argv[]) {
   }
   p_th = proper_primes[current_prime_pos];
 
+  InitConstants();
   InitBinomials();
   NTL::RR::SetPrecision(NUM_BITS_REAL_MANTISSA);
   pi = NTL::ComputePi_RR();
